Delete animation_resource.h
We're going to use a different class.

RE: 2D animations

At present, we only support sprite-atlas based animations, not world-space-transformational animations.  As a result, our "animations" are sequences of indices, which pertain specifically to a single sprite atlas, and we can presume we'll never use those animations without also using the assocated sprite atlas.

Therefore, for now, we're going to fold the animations into the atlas resource itself, and therefore, the class we make has to inheret from qRenderResource (which itself inherets from qResource).

We will inherret from qRenderResource, b/c the sprite atlas needs to be loaded in order to be show, and qRR already implements this functionality