gross bulk check in
1. unit tests all pass
2. smoketest passes

This is a gross, unstructured check in, but I want to get everything saved.  The embedded animations seem to be parsing correctly.

The sprite atlas displays, we just need to clip it down to the single index we want to show

Then we need to pass in a deltaTime, and have the atlas provide an updated source_rect, based on which animation is playing, the timing of the animation, and how long has passed since the last anim frame.

Note, all animations are allowed to have different timing, but within an animation, each frame  is required to keep the same timing.  You can fake it some, by specifying a frame multiple times in a row, in order for it to be long.  (And do it for everything except one to make that frame short)

For the moment, don't worry about overhead, this is just processing unsigned ints and floats, for the most part