[NOTE] Tests Pass, Smoketest Passes
this is a trivial change, made to enter a note into the depo:

at this time, the build looks "healthy", but I still haven't been able to get the sprites updating correctly.  On the plus side, we have made progress today, in particular regarding the math parts of calculating the source rectangle of a texture, based on knowledge of the textures, knowledge of the tile size, and the index of the tile to be rendered.

There's a break in the update chain somewhere though, will have to look for it after a break