Shader class code
1. this isn't commented out, but it's also not hooked up

2. In order to call the open gl fns, we have to have an opengl context.  The context isn't created during the unit test phase.  Our options are basically:

a. create an open GL context in unit test phase, and tear it back down before starting the main program

b. don't run unit tests on shader compilation, and rely on errors to come out correctly in logs)

I'm not sure what to do, so I'm gonna think about it for a while, then decide

3. other work:

a. OpenGLShaderProgram is split into small fns, and I feel good about it overall
b. class is starting to deviate from the tutorial code I was using to build it, which is also good