Update + decision
1. Update changes a few things about the code to make it feel better to me.

A. unique_ptr is used for the info buffer; we do this b/c unique_ptr will automagically dispose of the memory in the buffer.  Before, it was part of the stack frame, so this might actually be a "worse" implementation, but I still feel better about it right now
(if we wanted to transfer the buffer somewhere, we could)

B. changed Use() to SetActive().  This change is mostly just to help us diverge from the original code, so we're not copying them as much.

2. We can (I think) set up an OpenGL context "without" a window, as long as we don't try to render to the window.  In this case, we're just trying to compile our shaders, to make sure the error reporting works, if we have bad shaders.  Future work to do what I just said, but at present, we're not going to do it (even though it's a tech debt sprint), because in this first project there will likely just be a single set of basic shaders to get things on screen, which won't change during the project.

If we add shaders to the project, then we'll add tests at that time.