Update Render Commands
1. the renderer can also accept some invoke commands through the message bus.  Right now, it's just toggle wireframe rendering

2. supporting #1, updated engine class, so it has a send_render_command fn.

3. also supporting #1, added contains_term method to render manager.  This shadows a method by the same name, which console uses.  For now, I'm just gonna have a copy in every class that needs it, b/c I think only the manager classes will need it, and I don't want to add it to every game object there is.  (that can change in the future)

N. also, render manager requests a raw (non-owning) ptr to the debug console, using the message bus.  It uses this ptr to call the console's draw fn at the appropriate time, when rendering the scene.  Console is still in charge of formatting the render, and early-out-ing if it is meant to be hidden