Refactor Resource Manager - support for audio files
1. register_resource(unique_ptr<engine resource base> ptr)

a. refactored final part of load resource into fn #1.  Now, it's easier for an asset file to contain info for multiple resources, because we needn't return back out to the parse loop in order to register

2. refactored parse_file_version__010 to use #1

3. fn process_xml_audio_block(xml node)
a. added #3, to process individual sound asset files

4. refactor build_sound_resource_from_xml(xml node)
a. still returns unique_ptr to resource

5. general refactor, changed _name to _token, for variables which parse xml token names.  object names are still called name